+++ src/common/comp_tests/check_method_config.py
@@ -1,6 +1,5 @@
 import yaml
 
-
 ## VIASH START
 
 meta = {
@@ -29,7 +28,13 @@ def assert_dict(dict, functionality):
         for key in dict:
             assert key in arg_names or info, f"{key} is not a defined argument or .functionality.info field"
 
+def check_url(url):
+    import requests
 
+    get = requests.get(url)
+
+    assert get.status_code is (200 or 429), f"{url} is not reachable, {get.status_code}." # 429 rejected, too many requests
+        
 
 print("Load config data", flush=True)
 with open(meta["config"], "r") as file:
@@ -56,6 +61,9 @@ if ("control" not in info["type"]):
     assert "reference" in info, "reference not an info field"
     assert "documentation_url" in info is not None, "documentation_url not an info field or is empty"
     assert "repository_url" in info is not None, "repository_url not an info field or is empty"
+    check_url(info["documentation_url"])
+    check_url(info["repository_url"])
+
 
 if "variants" in info:
     for key in info["variants"]: